Move modules over from trade-tariff-platform
We've begun correcting trade-tariff-platform and have decided to move
those modules out of that repository in favor of a single
terraform-modules repository. In theory as this grows this isn't the
best way to go about seperating modules, we should eventually move to
using single repositories that we can test but this will be perfectly
fine for now.

This commit essentially just copy pasta's the code across from
trade-tariff-platform.

Signed-off-by: Chris Mills <7100370+chriscoffee@users.noreply.github.com>