BAU: Use new IAM roles in service module
We have some basic IAM roles but these need to be consumed by the tasks
that the module creates. I found some redundant configuration so I've
simplified where possible according to the AWS documentation.

I've added the output for both new roles so that we can pass in extra
policy attachments.