HMRC-964: Adds simple branch protection
This came out of a review of merge-queues and what they solve.

merge-queues add quite a lot more complexity to how we manage merging
feature branches and its actually significantly simpler to just make
sure the feature branch is up-to-date with the main branch before the
engineer merges/rebases their branch onto the merge-base.